// src/app/shared/models/input-type.enum.ts
export enum INPUT_TYPE_ENUM {
  TEXT = 'TEXT',
  TEXTAREA = 'TEXTAREA',
  NUMBER = 'NUMBER',
  OBJECT = 'OBJECT',
  ARRAY = 'ARRAY',
  BOOLEAN = 'BOOLEAN',
  DATE = 'DATE',
  DATETIME = 'DATETIME',
  IMAGE = 'IMAGE',
  PASSWORD = 'PASSWORD',
  SELECT = 'SELECT',


  // TEXT = 'TEXT',
  // NUMBER = 'NUMBER',
  // DATE = 'DATE',
  // TIME = 'TIME',
  // DATETIME = 'DATETIME',
  // SELECT = 'SELECT',
  // CHECKBOX = 'CHECKBOX',
  // RADIO = 'RADIO',
  // TEXTAREA = 'TEXTAREA',
  // FILE = 'FILE',
  // PASSWORD = 'PASSWORD',
  // EMAIL = 'EMAIL',
  // URL = 'URL',
  // COLOR = 'COLOR',
  // RANGE = 'RANGE',
  // SEARCH = 'SEARCH',
  // TEL = 'TEL',
  // HIDDEN = 'HIDDEN',
  // IMAGE = 'IMAGE',
  // MONTH = 'MONTH',
  // WEEK = 'WEEK',
  // BUTTON = 'BUTTON',
  // SUBMIT = 'SUBMIT',
  // TIMELOCAL = 'TIMELOCAL',
  // ARRAY = 'ARRAY',
}